requirements.txt
blinker
certifi
charset-normalizer
click
Flask
gunicorn
idna
itsdangerous
Jinja2
MarkupSafe
packaging
requests
urllib3
Werkzeug
BeautifulSoup4
python-dotenv